[workloadmeta/collectors/kubelet] Extract ParseKubeletPods
